requirements.txt
numpy==1.21.6
Pillow==9.0.1
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
imutils==0.5.4
einops==0.6.0
opencv-python==4.7.0.72
timm==0.5.4
mmcv==1.7.1
scipy==1.7.3
albumentations==1.3.0
tensorboardX==2.6
matplotlib==3.5.3
xformers==0.0.16
